% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{adjust_p_values}
\alias{adjust_p_values}
\title{Adjust p-values and return subsetted dataframe}
\usage{
adjust_p_values(resdf, alpha = 0.05)
}
\arguments{
\item{resdf}{A dataframe containing p-values in a column named 'p'.}

\item{alpha}{The significance level (default is 0.05).}
}
\value{
A subset of 'resdf' dataframe containing rows with significant p-values.
}
\description{
This function adjusts p-values using the BY, BH, or uncorrected methods and
returns a subset of the input dataframe where the adjusted p-values are significant.
}
\examples{
resdf <- data.frame(p = c(0.01, 0.05, 0.2, 0.3))
adjust_p_values(resdf)
}
