% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{antspymm_simlr}
\alias{antspymm_simlr}
\title{Perform SiMLR Analysis on Multimodal ANTsPyMM Data}
\usage{
antspymm_simlr(
  blaster,
  select_training_boolean,
  connect_cog,
  energy = c("cca", "reg", "lrr"),
  nsimlr = 5,
  covariates = "1",
  myseed = 3,
  doAsym = TRUE,
  returnidps = FALSE,
  restrictDFN = FALSE,
  doperm = FALSE
)
}
\arguments{
\item{blaster}{A dataframe containing multimodal data for analysis.}

\item{select_training_boolean}{boolean vector to define which entries are in training data}

\item{connect_cog}{Logical indicating whether cognitive data should be included. Defaults to FALSE.}

\item{energy}{The type of energy model to use for similarity analysis. Defaults to 'reg'.}

\item{nsimlr}{Number of similarity analyses to perform. Defaults to 5.}

\item{covariates}{any covariates to adjust training matrices}

\item{myseed}{Seed for random number generation to ensure reproducibility. Defaults to 3.}

\item{doAsym}{Logical indicating whether asymmetry predictors should be included. Defaults to TRUE.}

\item{returnidps}{Logical indicating whether to return the intermediate processing steps' results. Defaults to FALSE.}

\item{restrictDFN}{Logical indicating whether to restrict analysis to default network features. Defaults to FALSE.}

\item{doperm}{Logical indicating whether to perform permutation tests. Defaults to FALSE.  Will randomize image features in the training data and thus leads to "randomized" but still regularized projections.}
}
\value{
A list containing the results of the similarity analysis and related data.
}
\description{
This function processes multimodal data using SiMLR. It is designed
to be flexible, allowing for various preprocessing steps and analysis options. The analysis
can be adjusted through multiple parameters, offering control over the inclusion of certain
data types, permutation testing, and more.
}
\examples{
# Example usage:
result <- antspymm_simlr(dataframe)
}
