% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{assign_qc_ratings_NM2DMT}
\alias{assign_qc_ratings_NM2DMT}
\title{Assign Quality Control Ratings Based on Specific Criteria}
\usage{
assign_qc_ratings_NM2DMT(
  df,
  z_coord_col = "NM2DMT_NM_substantianigra_z_coordinate",
  volume_col = "NM2DMT_NM_volume_substantianigra",
  avg_col = "NM2DMT_NM_avg_substantianigra",
  sd_col = "NM2DMT_NM_sd",
  max_col = "NM2DMT_NM_max",
  lohi = c(0.3, 0.7),
  volume_threshold = 500,
  avg_threshold = 2500,
  sd_threshold = 50,
  max_threshold = 5500,
  verbose = FALSE
)
}
\arguments{
\item{df}{A data frame containing the dataset to be rated. No default value, must be provided by the user.}

\item{z_coord_col}{Name of the column for the Z-coordinate measurement. Default is "NM2DMT_NM_substantianigra_z_coordinate".}

\item{volume_col}{Name of the column for the volume measurement. Default is "NM2DMT_NM_volume_substantianigra".}

\item{avg_col}{Name of the column for the average measurement. Default is "NM2DMT_NM_avg_substantianigra".}

\item{sd_col}{Name of the column for the standard deviation measurement. Default is "NM2DMT_NM_sd".}

\item{max_col}{Name of the column for the maximum value measurement. Default is "NM2DMT_NM_max".}

\item{lohi}{A numeric vector of length 2 specifying the lower and upper thresholds for the Z-coordinate. Defaults to c(0.3, 0.7).}

\item{volume_threshold}{A numeric value specifying the threshold for the volume measurement. Default is 500.}

\item{avg_threshold}{A numeric value specifying the threshold for the average measurement. Default is 2500.}

\item{sd_threshold}{A numeric value specifying the threshold for the standard deviation measurement. Default is 50.}

\item{max_threshold}{A numeric value specifying the threshold for the maximum value measurement. Default is 5500.}

\item{verbose}{boolean}
}
\value{
The original data frame with additional columns containing the quality control ratings. 1=pass; 0=fail.
}
\description{
This function assigns quality control ratings to a dataset based on several
criteria related to measurements of a specific anatomical structure, such as the substantia nigra.
The ratings are determined by whether the measurements fall within certain thresholds.
}
\examples{
# Example usage:
# df_rated <- assign_qc_ratings(df)
# View(df_rated)
}
