% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{asymind}
\alias{asymind}
\title{Asymmetry index for asymmetry analysis}
\usage{
asymind(left, right)
}
\arguments{
\item{left}{Numeric. The value representing the left side of the measurement.}

\item{right}{Numeric. The value representing the right side of the measurement.}
}
\value{
Numeric value representing the Symmetry-Adjusted Mean (SAM).
}
\description{
The formula is:
\deqn{asymind = \frac{2|A-B|}{(A+B)} }
}
\examples{
# Example usage:
left <- 5
right <- 3
asymind(left, right)

# For a symmetric case:
asymind(5, 5)

}
