% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{asymmetry}
\alias{asymmetry}
\title{Compute Various Asymmetry Metrics Between Two Sides}
\usage{
asymmetry(left, right, name_of_measurement)
}
\arguments{
\item{left}{Numeric. The value representing the left side of the measurement.}

\item{right}{Numeric. The value representing the right side of the measurement.}

\item{name_of_measurement}{Character. The name of the asymmetry measure to compute. Options include:
\itemize{
\item "absolute_difference": Absolute difference between left and right.
\item "proportional_asymmetry": Proportional asymmetry.
\item "symmetry_adjusted_mean": Symmetry-adjusted mean (SAM).
\item "relative_asymmetry": (Left - Right) / (Left + Right).
\item "asymmetry_index": (|Left - Right|) / ((Left + Right) / 2).
}}
}
\value{
Numeric value of the computed asymmetry measure.
}
\description{
This function computes different types of asymmetry metrics given left and right side measurements.
It can return standard asymmetry indices, including the absolute difference, symmetry-adjusted mean,
proportional asymmetry, and more.
}
\examples{
asymmetry(5, 3, "absolute_difference")
asymmetry(10, 8, "symmetry_adjusted_mean")

}
