% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{augment_with_custom_color}
\alias{augment_with_custom_color}
\title{Augment a Data Frame with Custom Color Column}
\usage{
augment_with_custom_color(
  df,
  column_name,
  color_palette,
  color_palette_name = NULL
)
}
\arguments{
\item{df}{A data frame to be augmented.}

\item{column_name}{The name of the column in \code{df} whose factor levels are to be
mapped to colors. This column should exist in \code{df} and should be a factor or
convertible to a factor.}

\item{color_palette}{A vector of colors, corresponding to the factor levels of
the specified column. The number of colors must match the number of unique
factor levels in the column.}

\item{color_palette_name}{optional named color palette}
}
\value{
A data frame identical to \code{df} but with an additional column
'custom_color', which contains the color mappings.
}
\description{
This function adds a new column named 'custom_color' to a given data frame.
The new column maps factor levels of a specified column to colors
from a user-defined palette.
}
\examples{
data <- data.frame(
  category = c("A", "B", "C", "A", "B"),
  value = c(10, 20, 30, 40, 50)
)
palette <- c("red", "green", "blue")
augmented_data <- augment_with_custom_color(data, "category", palette)

}
