% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{averageRepeats}
\alias{averageRepeats}
\title{Average repeat data}
\usage{
averageRepeats(mxdfin, idvar, visitvar)
}
\arguments{
\item{mxdfin}{Input data frame with repeated measurements and a grouped time variable}

\item{idvar}{variable name for unique subject identifier column}

\item{visitvar}{variable name for the visit or date column}
}
\value{
data frame
}
\description{
Average repeat data
}
\examples{
mydf = generateSubtyperData( 100 )
mydfhq = averageRepeats( mydf, "Id", "visit")
}
\author{
Avants BB
}
