% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{balanceDataMultiClass}
\alias{balanceDataMultiClass}
\title{balanced sampling of a multi-class variable}
\usage{
balanceDataMultiClass(x, variable, method)
}
\arguments{
\item{x}{data frame to extend}

\item{variable}{column name}

\item{method}{one of permissible methods listed in function (will print if wrong method passed)}
}
\value{
new data frame
}
\description{
resample a dataframe to counteract variable imbalance
}
\author{
Avants BB
}
