% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{balancedDataframe}
\alias{balancedDataframe}
\title{balanced sampling of a variable}
\usage{
balancedDataframe(x, variable, method)
}
\arguments{
\item{x}{data frame to extend}

\item{variable}{column name}

\item{method}{one of}
}
\value{
new data frame
}
\description{
resample a dataframe to counteract variable imbalance
}
\author{
Avants BB
}
