% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{build_api_config}
\alias{build_api_config}
\title{Build API config (URL, model, key)}
\usage{
build_api_config(
  backend,
  api_key_env = NULL,
  model = NULL,
  skip_api_key_check = FALSE
)
}
\description{
Build API config (URL, model, key)
}
