% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{cache_write}
\alias{cache_write}
\title{Cache Write Helper}
\usage{
cache_write(cache_dir, key, parsed)
}
\arguments{
\item{cache_dir}{The directory to use for caching.}

\item{key}{The cache key, which will form the filename.}

\item{parsed}{The object to be serialized to JSON and cached.}
}
\description{
Writes an object as a JSON file to the specified cache directory.
}
