% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{clearcolname}
\alias{clearcolname}
\title{clearcolname}
\usage{
clearcolname(mydf, mycolname)
}
\arguments{
\item{mydf}{dataframe input}

\item{mycolname}{columns for the task}
}
\value{
trimmed dataframe
}
\description{
remove a column from a dataframe
}
\author{
Avants BB
}
