% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{compare_model_predictions}
\alias{compare_model_predictions}
\title{Compare Model Predictions with True Values}
\usage{
compare_model_predictions(model1, model2, data, response, re.form = NULL)
}
\arguments{
\item{model1}{A fitted model object (either a standard regression or a mixed model).}

\item{model2}{A second fitted model object (either a standard regression or a mixed model).}

\item{data}{A data frame containing the variables used in the models.}

\item{response}{The name of the response variable as a string.}

\item{re.form}{A formula specifying which random effects to include in predictions, or \code{NA} to exclude them. Defaults to including all random effects. Only applies if the model is a mixed model.}
}
\value{
A list of two ggplot objects, each comparing the predictions of a model with the true response variable.
}
\description{
This function takes in two models and produces two ggplots in a list that show the predictions
of the models versus the true response variable. For mixed models, it allows the user to ignore random effects.
}
\examples{
\dontrun{
  mdl1 <- lm(Sepal.Length ~ Sepal.Width + Petal.Length, data = iris)
  mdl2 <- lm(Sepal.Length ~ Sepal.Width + Petal.Width, data = iris)
  plots <- compare_model_predictions(mdl1, mdl2, data = iris, response = "Sepal.Length")
  plots[[1]] # Plot for model 1
  plots[[2]] # Plot for model 2
}
}
