% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{complete_cases_from_equation}
\alias{complete_cases_from_equation}
\title{Extract Complete Cases from a Data Frame Based on a Model Equation}
\usage{
complete_cases_from_equation(df, equation)
}
\arguments{
\item{df}{A data frame containing the variables specified in the equation.}

\item{equation}{A model equation specifying the variables of interest.}
}
\value{
A boolean vector indicating which rows of the data frame have complete cases.
}
\description{
This function takes a data frame and a model equation as input, and returns
a boolean vector indicating which rows of the data frame have complete cases
for the variables specified in the equation.
}
\examples{
df <- data.frame(
  x = c(1, 2, NA, 4),
  y = c(NA, 2, 3, 4),
  z = c(1, 2, 3, NA)
)
equation <- y ~ x + z
complete_index <- complete_cases_from_equation(df, equation)
print(complete_index)  # Output: FALSE TRUE FALSE FALSE
}
