% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{consensusSubtypingCOCA}
\alias{consensusSubtypingCOCA}
\title{consensusSubtypingCOCA}
\usage{
consensusSubtypingCOCA(
  dataToClust,
  targetk,
  cocanames,
  newclustername,
  reorderingVariable,
  idvar,
  visitName,
  baselineVisit,
  maxK,
  consensusmethod = "kmeans",
  returnonehot = FALSE,
  binnmf = 0,
  verbose = TRUE
)
}
\arguments{
\item{dataToClust}{dataframe input that contains the relevant variables (may have others as well)}

\item{targetk}{the desired number of classes}

\item{cocanames}{names of columns to use for the consensus}

\item{newclustername}{the column name for the consensus clustering}

\item{reorderingVariable}{the name of the column to use to reorder the cluster names}

\item{idvar}{variable name for unique subject identifier column}

\item{visitName}{the column name defining the visit variables}

\item{baselineVisit}{the string naming the baseline visit}

\item{maxK}{maximum number of clusters}

\item{consensusmethod}{either kmeans or hclust}

\item{returnonehot}{boolean}

\item{binnmf}{integer (0 is default - no nmf)}

\item{verbose}{boolean}
}
\value{
new dataframe with new variables attached
}
\description{
apply consensus clustering give several clustering solutions
}
\examples{
mydf = generateSubtyperData( 100 )
}
\author{
Avants BB
}
