% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{consensusSubtypingPredict}
\alias{consensusSubtypingPredict}
\title{consensusSubtypingPredict}
\usage{
consensusSubtypingPredict(
  dataToClust,
  featureNames,
  clustVec,
  clustmodels,
  reorderers,
  mvcl,
  idvar,
  visitName,
  baselineVisit
)
}
\arguments{
\item{dataToClust}{dataframe input that contains the relevant variables (may have others as well)}

\item{featureNames}{names to use in the clustering}

\item{clustVec}{names of the clustering methods to use}

\item{clustmodels}{the trained models}

\item{reorderers}{the reordering data frames}

\item{mvcl}{character prefix for the new cluster column names}

\item{idvar}{variable name for unique subject identifier column}

\item{visitName}{the column name defining the visit variables}

\item{baselineVisit}{the string naming the baseline visit}
}
\value{
new dataframe with new variables attached
}
\description{
apply several clustering methods and append results to a dataframe;
assumes trained models already exist.
}
\examples{
mydf = generateSubtyperData( 100 )
}
\author{
Avants BB
}
