% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{consensusSubtypingPrep}
\alias{consensusSubtypingPrep}
\title{consensusSubtypingPrep}
\usage{
consensusSubtypingPrep(
  dataToTrain,
  dataToPredict,
  featureNames,
  clustVec,
  maxK,
  reorderingVariable,
  mvcl,
  idvar,
  visitName,
  baselineVisit,
  whichrank = 0,
  ntoreturnperk = 2,
  verbose = FALSE
)
}
\arguments{
\item{dataToTrain}{dataframe input that contains the relevant variables (may have others as well) on which training will be based}

\item{dataToPredict}{dataframe input that contains the relevant variables (may have others as well) for which prediction will be done}

\item{featureNames}{names to use in the clustering}

\item{clustVec}{names of the clustering methods to use}

\item{maxK}{the maximum desired number of classes}

\item{reorderingVariable}{the name of the column to use to reorder the cluster names}

\item{mvcl}{character prefix for the new cluster column names}

\item{idvar}{variable name for unique subject identifier column}

\item{visitName}{the column name defining the visit variables}

\item{baselineVisit}{the string naming the baseline visit}

\item{whichrank}{allows user to get 2nd (or 3rd) rank set of methods}

\item{ntoreturnperk}{number of method results per k to return}

\item{verbose}{boolean}
}
\value{
new dataframe with new variables attached
}
\description{
generate input for consensus clustering give several clustering algorithms
}
\examples{
mydf = generateSubtyperData( 100 )
}
\author{
Avants BB
}
