% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{consensusSubtypingTrain}
\alias{consensusSubtypingTrain}
\title{consensusSubtypingTrain}
\usage{
consensusSubtypingTrain(
  dataToClust,
  featureNames,
  clustVec,
  ktrain,
  reorderingVariable,
  mvcl = "MVST",
  verbose = FALSE
)
}
\arguments{
\item{dataToClust}{dataframe input that contains the relevant variables (may have others as well)}

\item{featureNames}{names to use in the clustering}

\item{clustVec}{names of the clustering methods to use}

\item{ktrain}{the number of clusters}

\item{reorderingVariable}{the name of the column to use to reorder the cluster names}

\item{mvcl}{character prefix for the new cluster column names}

\item{verbose}{boolean}
}
\value{
a list with newdata: dataframe with new variables attached; models contains the trained models; reorderers contains a dataframe for reordering cluster levels
}
\description{
apply several clustering methods and append results to a dataframe;
usually will include one subject per row
}
\examples{
mydf = generateSubtyperData( 100 )
}
\author{
Avants BB
}
