% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{convert_to_random_effects}
\alias{convert_to_random_effects}
\title{Convert Character Vector to Random Effects Terms for lme4}
\usage{
convert_to_random_effects(variables)
}
\arguments{
\item{variables}{A character vector where each element is the name of a grouping factor
for which a random intercept will be included in the model.}
}
\value{
A character vector of random effect terms, each formatted as \code{(1|group)},
where \code{group} is replaced with each string from the input vector.
}
\description{
This function takes a vector of strings representing grouping factors and converts
each string into a random effect term formatted for use in lme4 model formulas.
Specifically, it formats them as random intercepts for the specified grouping factors.
}
\examples{
grouping_factors <- c("school", "class")
random_effects <- convert_to_random_effects(grouping_factors)
print(random_effects)

}
