% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{count_unique_subjects_per_diagnosis}
\alias{count_unique_subjects_per_diagnosis}
\title{Count Unique Subjects per Diagnosis}
\usage{
count_unique_subjects_per_diagnosis(
  df,
  diagnosis_col,
  subject_col,
  format_counts = FALSE
)
}
\arguments{
\item{df}{A dataframe containing the data.}

\item{diagnosis_col}{A string specifying the column name for diagnoses.}

\item{subject_col}{A string specifying the column name for subjects.}

\item{format_counts}{A boolean indicating whether to format counts as a single string
in the format "x/y/z/q" where x, y, z, q represent the counts for each diagnosis. Defaults to FALSE.}
}
\value{
A dataframe with two columns: Diagnosis and Unique_Subjects. If \code{format_counts} is TRUE,
only the formatted counts are returned.
}
\description{
This function counts the number of unique subjects for each diagnosis in a dataframe.
If \code{format_counts} is \code{TRUE}, it returns a formatted string of counts.
}
\examples{
\dontrun{
df <- data.frame(
  diagnosis = c('A', 'A', 'B', 'A', 'B', 'C', 'C', 'C'),
  subject_id = c(1, 2, 2, 3, 4, 5, 6, 6)
)
count_unique_subjects_per_diagnosis(df, 'diagnosis', 'subject_id')
count_unique_subjects_per_diagnosis(df, 'diagnosis', 'subject_id', format_counts = TRUE)
}
}
