% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{create_correlated_vector}
\alias{create_correlated_vector}
\title{Create a new vector with a specified correlation to an input vector}
\usage{
create_correlated_vector(a, rho)
}
\arguments{
\item{a}{Input vector}

\item{rho}{Desired correlation between a and the new vector}
}
\value{
A new vector with the desired correlation to a
}
\description{
Create a new vector with a specified correlation to an input vector
}
\examples{
a <- rnorm(100)
rho <- 0.7
b <- create_correlated_vector(a, rho)
cor(a, b)
}
