% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{create_radar_chart}
\alias{create_radar_chart}
\title{Create a Radar Chart}
\usage{
create_radar_chart(data, group_color, meansd = TRUE)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{group_color}{A named vector of colors to be used for each group in the radar chart.}

\item{meansd}{A logical value indicating whether to use mean + 2SD and mean as range (default is FALSE).}
}
\value{
A radar chart plot.
}
\description{
This function creates a radar chart for the specified data.
}
\examples{
# group_color <- c("Group 1" = "red", "Group 2" = "blue", "Group 3" = "green")
# create_radar_chart(data, group_color)
}
