% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{dcurvarsel}
\alias{dcurvarsel}
\title{dcurvarsel}
\usage{
dcurvarsel(curdf, variablenames, fraction)
}
\arguments{
\item{curdf}{dataframe input}

\item{variablenames}{columns for the task}

\item{fraction}{scalar between zero and one}
}
\value{
vector of selected variables
}
\description{
variable selection with CUR - using default parameters from package example
}
\examples{
mydf = generateSubtyperData( 100 )
vtosel = getNamesFromDataframe("Random",mydf)
# vimp = dcurvarsel( mydf, vtosel, 0.5 )
}
\author{
Avants BB
}
