% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{decode_antspymm_label}
\alias{decode_antspymm_label}
\title{Decode an ANTsPyMM label into a structured 4-column format.}
\usage{
decode_antspymm_label(label)
}
\arguments{
\item{label}{\code{character(1)} The raw label string.}
}
\value{
A named list: \code{modality}, \code{laterality}, \code{measurement}, \code{anatomy}.
}
\description{
Decode an ANTsPyMM label into a structured 4-column format.
}
