% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{delta_R2_to_f2}
\alias{delta_R2_to_f2}
\title{Calculate Cohen's f-squared from Delta R-squared}
\usage{
delta_R2_to_f2(delta_r2, r2_full)
}
\arguments{
\item{delta_r2}{The change in R-squared (numeric) when the predictors of
interest are added to the model (i.e., R-squared of the full model minus
R-squared of the base model).}

\item{r2_full}{The R-squared (numeric) of the full model that includes the
predictors of interest.}
}
\value{
The numeric value for Cohen's f-squared.
}
\description{
Calculates the Cohen's f-squared effect size, a measure of local effect size
for a set of predictors in a hierarchical regression model.
}
\examples{
# Calculate f-squared for a set of predictors that added 5\% to the variance
# explained, where the final model explained 30\% of the variance.
delta_R2_to_f2(delta_r2 = 0.05, r2_full = 0.30)
}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences}
(2nd ed.). Hillsdale, NJ: Erlbaum.
}
