% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{.decode_connectivity_label}
\alias{.decode_connectivity_label}
\title{Decodes a connectivity label into its four components.}
\usage{
.decode_connectivity_label(cleaned_label)
}
\description{
Decodes a connectivity label into its four components.
}
\keyword{internal}
