% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{.decode_standard_label}
\alias{.decode_standard_label}
\title{Decodes a standard (non-connectivity) label into its four components.}
\usage{
.decode_standard_label(cleaned_label)
}
\description{
Decodes a standard (non-connectivity) label into its four components.
}
\keyword{internal}
