% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{.get_master_dictionary}
\alias{.get_master_dictionary}
\title{Master Dictionary for ANTsPyMM Anatomical and File-Type Labels}
\usage{
.get_master_dictionary()
}
\description{
Master Dictionary for ANTsPyMM Anatomical and File-Type Labels
}
\keyword{internal}
