% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{.preprocess_label}
\alias{.preprocess_label}
\title{Pre-processes a raw label by expanding abbreviations and normalizing it.}
\usage{
.preprocess_label(label)
}
\description{
Pre-processes a raw label by expanding abbreviations and normalizing it.
}
\keyword{internal}
