% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{fill1col2another}
\alias{fill1col2another}
\title{fill1col2another}
\usage{
fill1col2another(df, x, y)
}
\arguments{
\item{df}{dataframe input}

\item{x}{column to fill}

\item{y}{column to fill from}
}
\value{
filled dataframe
}
\description{
fill missing data values into one column from another column;
will only fill values that are NA
}
\author{
Avants BB
}
