% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{filterNAcolumns}
\alias{filterNAcolumns}
\title{Filter Columns by Percentage of NA Values}
\usage{
filterNAcolumns(df, percentThreshold)
}
\arguments{
\item{df}{A dataframe from which columns will be filtered.}

\item{percentThreshold}{A numeric value between 0 and 100 indicating the percentage
threshold of NA values. Columns with NA percentages greater or equal to this
value will be removed from the dataframe.}
}
\value{
A dataframe with columns filtered based on the NA percentage threshold.
}
\description{
This function filters out columns from a dataframe that have a percentage of
NA (missing) values greater or equal to a user-defined threshold. The percentage
is calculated with respect to the total number of rows in the dataframe.
}
\examples{
data(mtcars)
mtcars[1:5, 1:2] <- NA # Introduce NA values for demonstration
filtered_df <- filterNAcolumns(mtcars, 10)
print(filtered_df)

}
