% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{find_closest_subjects}
\alias{find_closest_subjects}
\title{Find Closest Subjects}
\usage{
find_closest_subjects(reference_df, target_df, k, sex_col_name, age_col_name)
}
\arguments{
\item{reference_df}{A data frame containing the reference dataset.}

\item{target_df}{A data frame containing the target dataset to search within.}

\item{k}{The number of closest subjects to find.}

\item{sex_col_name}{The name of the column in the datasets that contains subjects' sex.}

\item{age_col_name}{The name of the column in the datasets that contains subjects' age.}
}
\value{
A data frame with the rows from \code{target_df} that are closest to \code{reference_df} based on the criteria.
}
\description{
This function identifies the closest subjects in a target dataset to a reference dataset
based on specified criteria such as sex and age.
}
\examples{
# find_closest_subjects(referenceData, targetData, 5, "Gender", "Age")
}
