% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{force_unload_package}
\alias{force_unload_package}
\title{Forcefully Unload a Package in R}
\usage{
force_unload_package(pkg)
}
\arguments{
\item{pkg}{The name of the package to unload as a string.}
}
\value{
Logical TRUE if the package was successfully unloaded, FALSE otherwise.
}
\description{
This function forcefully unloads a package from R by detaching it from the search path,
unloading its namespace, and trying to remove it from all environments.
}
\examples{
\dontrun{
  force_unload_package("ggplot2")
}
}
