% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{generateSubtyperData}
\alias{generateSubtyperData}
\title{Generate example data for illustrating \code{subtyper}}
\usage{
generateSubtyperData(n = 100)
}
\arguments{
\item{n}{numer of subjects}
}
\value{
data frame
}
\description{
Generate example data for illustrating \code{subtyper}
}
\examples{

mydf = generateSubtyperData( 5 )

}
\author{
Avants BB
}
