% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{getExt}
\alias{getExt}
\title{Get the file extension from a file-name. from \code{reader} package.}
\usage{
getExt(fn)
}
\arguments{
\item{fn}{filename(s) (with full path is ok too)}
}
\value{
returns the (usually) 3 character file extension of a filename
}
\description{
Get the file extension from a file-name. from \code{reader} package.
}
\author{
Nicholas Cooper \email{nick.cooper@cimr.cam.ac.uk}
}
