% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{getNamesFromDataframe}
\alias{getNamesFromDataframe}
\title{Extract column names with concatenated search parameters}
\usage{
getNamesFromDataframe(x, demogIn)
}
\arguments{
\item{x}{vector of strings}

\item{demogIn}{the dataframe with column names to search.}
}
\value{
vector of string column names
}
\description{
Extract column names with concatenated search parameters
}
\examples{

mydf = generateSubtyperData( 5 )
nms = getNamesFromDataframe( c("it","v"), mydf )

}
\author{
Avants BB
}
