% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{kable_table}
\alias{kable_table}
\title{Create a Styled LaTeX Table with Adjustable Size and Orientation}
\usage{
kable_table(
  data,
  caption,
  scl = 0.75,
  row.names = FALSE,
  striped = TRUE,
  landscape = TRUE,
  table_size = "medium",
  digits = 3,
  latex_options = c("striped"),
  format = "latex"
)
}
\arguments{
\item{data}{A data frame to be displayed in the table.}

\item{caption}{A character string for the table caption.}

\item{scl}{A numeric value to scale the table size. Defaults to 0.75.}

\item{row.names}{Logical, whether to include row names in the table. Defaults to FALSE.}

\item{striped}{Logical, whether to apply striped styling to the table. Defaults to TRUE.}

\item{landscape}{Logical, whether to rotate the table to landscape mode. Defaults to TRUE.}

\item{table_size}{A character string for adjusting the overall size of the table. Options include "small", "medium", "large". Defaults to "medium".}

\item{digits}{An integer specifying the number of decimal places for numeric values. Defaults to 3.}

\item{latex_options}{A character vector for specifying additional LaTeX styling options. Defaults to \code{c("striped")}.}

\item{format}{either latex or html}
}
\value{
A LaTeX table generated with \code{kableExtra}, ready for inclusion in an RMarkdown or LaTeX document.
}
\description{
This function generates a LaTeX table from a data frame using the \code{kableExtra} package,
with options to adjust the table size, apply striped styling, and set the table orientation
to landscape or portrait mode.
}
\examples{
data <- mtcars[1:5, 1:5]
kable_table(data, caption = "Sample Table", scl = 0.8, row.names = TRUE, striped = TRUE, landscape = FALSE, table_size = "large")

}
