% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{lm_anv_p_and_d}
\alias{lm_anv_p_and_d}
\title{Fit and Compare Linear Models}
\usage{
lm_anv_p_and_d(
  data,
  outcome,
  predictor,
  covariates,
  predictoroperator = "*",
  verbose = FALSE,
  plot_interaction_with = NULL,
  interact_with_all = FALSE
)
}
\arguments{
\item{data}{A data frame containing all necessary variables.}

\item{outcome}{The outcome variable name (character string).}

\item{predictor}{A character string specifying the main predictor(s).}

\item{covariates}{A character string of fixed-effect covariates (e.g., "Age + Sex").}

\item{predictoroperator}{Operator for combining predictors. Defaults to \code{'+'}.
Use \code{'*'} to model full interactions.}

\item{verbose}{A logical. If \code{TRUE}, the model formulas used in each run will be printed.}

\item{plot_interaction_with}{A moderator variable name to plot as an interaction.
When \code{predictoroperator = "*"}, this also specifies the \emph{only} covariate
that will be included in the interaction term, unless \code{interact_with_all} is \code{TRUE}.}

\item{interact_with_all}{A logical. If \code{FALSE} (default) and \code{predictoroperator = "*"},
only the variable in \code{plot_interaction_with} is interacted with the predictors.
If \code{TRUE}, all \code{covariates} are interacted with the predictors.}
}
\value{
A harmonized list of model results.
}
\description{
Internal helper for \code{lm} models with flexible interaction specification.
}
