% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{lmer_anv_p_and_d}
\alias{lmer_anv_p_and_d}
\title{Fit and compare linear mixed-effects models and calculate effect sizes.}
\usage{
lmer_anv_p_and_d(
  data,
  outcome,
  predictor,
  fixed_effects,
  random_effects,
  predictoroperator = "*",
  verbose = FALSE
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{outcome}{A character string specifying the outcome variable.}

\item{predictor}{A character string specifying the main predictor of interest.}

\item{fixed_effects}{A character string of other fixed effects.}

\item{random_effects}{A character string specifying the random effects structure (e.g., "(1 | commonID)").}

\item{predictoroperator}{A character string for the operator between fixed effects and the predictor (e.g., '*' or '+').}

\item{verbose}{A logical indicating whether to print debugging information.}
}
\value{
A list containing the model, comparison, effect sizes, coefficients, and sample size, or NULL on error.
}
\description{
Fit and compare linear mixed-effects models and calculate effect sizes.
}
