% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{log_parameters}
\alias{log_parameters}
\title{Log Parameters of a Function Call}
\usage{
log_parameters(func, logfile, ...)
}
\arguments{
\item{func}{The function to be called and logged.}

\item{logfile}{The name of the log file where the function call information will be saved. The actual file will be named \code{logfile_function_calls.log}.}

\item{...}{Additional arguments to be passed to the function \code{func}.}
}
\value{
The result of the function \code{func}.
}
\description{
This function logs the parameters and the call time of a specified function to a log file.
If any of the parameters are matrices or vectors longer than 5 elements, their values will not be printed in the log.
}
\examples{
my_function <- function(a, b, c = 10) {
  return(a + b + c)
}
result <- log_parameters(my_function, "my_log", a = 1, b = 2, c = 3)
print(result)

}
