% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{mapAsymVar}
\alias{mapAsymVar}
\title{Convert left/right variables to a measure of asymmetry}
\usage{
mapAsymVar(mydataframe, leftvar, replacer = "Asym")
}
\arguments{
\item{mydataframe}{dataframe containing relevant variables}

\item{leftvar}{left variable names}

\item{replacer}{string to replace left with in column names of output}
}
\value{
fixed x
}
\description{
Convert left/right variables to a measure of asymmetry
}
\author{
Avants BB
}
