% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{mapLRAverageVar}
\alias{mapLRAverageVar}
\title{Convert left/right variables to an average measurement}
\usage{
mapLRAverageVar(mydataframe, leftvar, replacer = "LRAVG")
}
\arguments{
\item{mydataframe}{dataframe containing relevant variables}

\item{leftvar}{left variable names}

\item{replacer}{string to replace left with in column names of output}
}
\value{
fixed x
}
\description{
Convert left/right variables to an average measurement
}
\author{
Avants BB
}
