% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{mlr3classification}
\alias{mlr3classification}
\title{mlr3classification}
\usage{
mlr3classification(
  dfin,
  tcols,
  nrepeats = 10,
  partrate = 0.8,
  dup_size = 0,
  balancing = "smote",
  verbose = TRUE
)
}
\arguments{
\item{dfin}{dataframe input}

\item{tcols}{columns for the prediction task - first is the target outcome}

\item{nrepeats}{number of subsampling-driven train test runs}

\item{partrate}{partition ratio for the training 0.8 equals 80 percent train 20 test}

\item{dup_size}{integer for over/under/smote sampling}

\item{balancing}{string over, under, smote, none are the options}

\item{verbose}{boolean}
}
\value{
dataframe quantifying performance
}
\description{
cross-validate a mlr3 classification model
}
\author{
Avants BB
}
