% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{mlr3classifiercv}
\alias{mlr3classifiercv}
\title{mlr3classifiercv}
\usage{
mlr3classifiercv(
  dfin,
  tcols,
  nrepeats = 10,
  partrate = 0.8,
  dup_size = 0,
  balancing = "smote",
  mylearners = mlr3classifiers(),
  subjectIDs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dfin}{dataframe input}

\item{tcols}{columns for the prediction task - first is the target outcome}

\item{nrepeats}{number of subsampling-driven train test runs}

\item{partrate}{partition ratio for the training 0.8 equals 80 percent train 20 test}

\item{dup_size}{integer for over/under/smote sampling}

\item{balancing}{string over, under, smote, none are the options}

\item{mylearners}{the mlr3 learners over which to search ; defaults to those that support multiclass}

\item{subjectIDs}{unique IDs per subject; aids with repeated measurement data
by ensuring subjects exist in uniquely in either train or test}

\item{verbose}{boolean}
}
\value{
dataframe quantifying performance
}
\description{
cross-validate a mlr3 classification model
}
\author{
Avants BB
}
