% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{mlr3classifiers}
\alias{mlr3classifiers}
\title{mlr3classifiers}
\usage{
mlr3classifiers(twoclass = TRUE, all = FALSE)
}
\arguments{
\item{twoclass}{boolean}

\item{all}{boolean}
}
\value{
vector of strings
}
\description{
good classification learners from mlr3 (in my experience)
}
\author{
Avants BB
}
