% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{multigrep}
\alias{multigrep}
\title{Grep entries with a vector search parameters}
\usage{
multigrep(x, desc)
}
\arguments{
\item{x}{vector}

\item{desc}{a vector of search terms}
}
\value{
result of grep
}
\description{
Grep entries with a vector search parameters
}
\author{
Avants BB
}
