% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{nrgDateToRDate}
\alias{nrgDateToRDate}
\title{Convert nrg format date to R format date}
\usage{
nrgDateToRDate(x)
}
\arguments{
\item{x}{nrg format date}
}
\value{
fixed x
}
\description{
Convert nrg format date to R format date
}
\author{
Avants BB
}
