% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{outlierness}
\alias{outlierness}
\title{Outlierness scoring}
\usage{
outlierness(mxdfin, measureColumns, calck, report = FALSE)
}
\arguments{
\item{mxdfin}{Input data frame with repeated measurements and a grouped time variable}

\item{measureColumns}{vector string of column names to use in outlier detection}

\item{calck}{optional integer for knn}

\item{report}{boolean return vector string of column names for outlier detection results}
}
\value{
data frame
}
\description{
Produce several complementary measurements of outlierness
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
mydfol = outlierness( mydf, rbfnames )
}
\author{
Avants BB
}
