% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{oversample_minority}
\alias{oversample_minority}
\title{Oversample a minority class in a data frame}
\usage{
oversample_minority(x, vname)
}
\arguments{
\item{x}{data frame containing the data to be oversampled}

\item{vname}{character string specifying the name of the column containing the class labels}
}
\value{
data frame with the minority class oversampled to match the size of the majority class
}
\description{
This function oversamples the minority class in a data frame to match the size of the majority class.
}
\examples{
data <- data.frame(class = c(rep("A", 10), rep("B", 100)), 
                   feature1 = rnorm(110), 
                   feature2 = rnorm(110))
oversampled_data <- oversample_minority(data, "class")
}
