% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{plinkVariantsDataFrame}
\alias{plinkVariantsDataFrame}
\title{genetic variants data frame from plink data}
\usage{
plinkVariantsDataFrame(rootFileName, targetSNPs, verbose = FALSE)
}
\arguments{
\item{rootFileName}{root for pgen psam and pvar files}

\item{targetSNPs}{snps to extract (optional - if absent, will get all)}

\item{verbose}{boolean}
}
\value{
dataframes with both variants and subject ids
}
\description{
genetic variants data frame from plink data
}
\examples{
# mydf = plinkVariantsDataFrame( fn, c( 'rs6469804', 'rs6859' ) )
}
\author{
Avants BB
}
