% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{plotCategoricalData}
\alias{plotCategoricalData}
\title{Plot Categorical Data}
\usage{
plotCategoricalData(datatoplot, columnName)
}
\arguments{
\item{datatoplot}{A data frame containing the dataset to be plotted.}

\item{columnName}{The name of the column in the dataset that contains categorical data.}
}
\value{
A ggplot object representing the frequency of each category in the specified column.
}
\description{
This function creates a visualization for categorical data using ggplot2. It takes a dataset and a column name,
generates a frequency table for the categorical data, and plots the frequencies.
}
\examples{
# plotCategoricalData(myData, "myCategoryColumn")
}
