% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{predictSubtypeClusterMulti}
\alias{predictSubtypeClusterMulti}
\title{Predict subtype from multivariate data}
\usage{
predictSubtypeClusterMulti(
  mxdfin,
  measureColumns,
  clusteringObject,
  clustername = "GMMClusters"
)
}
\arguments{
\item{mxdfin}{Input data frame}

\item{measureColumns}{vector defining the data columns to be used for clustering.}

\item{clusteringObject}{a GMM object to predict clusters}

\item{clustername}{column name for the identified clusters}
}
\value{
the clusters attached to the data frame
}
\description{
This is the inference module for subtype definition based on a matrix.
Currently only supports clustering based on gaussian mixtures
via the ClusterR package.
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
gmmcl = trainSubtypeClusterMulti( mydf, rbfnames, maxk=4 )
gmmclp = predictSubtypeClusterMulti( mydf, rbfnames, gmmcl )
}
\author{
Avants BB
}
