% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{predictSubtypeUni}
\alias{predictSubtypeUni}
\title{Predict subtype for univariate data}
\usage{
predictSubtypeUni(mxdfin, subtypeDataFrame)
}
\arguments{
\item{mxdfin}{Input data frame}

\item{subtypeDataFrame}{data frame defining the subtypes and cutpoints}
}
\value{
data frame with attached subtypess
}
\description{
This is the inference module for subtype definition based on a vector.
After training, we can predict subtype very easily in new data based on
the data frame produced by training \code{trainSubtypeUni}.
}
\examples{
mydf = generateSubtyperData( 100 )
}
\author{
Avants BB
}
