% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{quantSquared}
\alias{quantSquared}
\title{quantify by quantiles (quantSquared)}
\usage{
quantSquared(vecToBeTransformed, matrixReferenceDistribution)
}
\arguments{
\item{vecToBeTransformed}{Input vector should have same entries as reference;
should be a row of a data frame}

\item{matrixReferenceDistribution}{data frame defining the reference data;
ideally this will not contain missing data.}
}
\value{
the quantile transformed vector
}
\description{
transform input vector to a quantile representation relative to a reference.
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
zz=data.frame(t(colMeans(mydf[,rbfnames])))
newvec = quantSquared( zz, mydf )
newvec2 = quantSquared( mydf[1,rbfnames], mydf )
}
\author{
Avants BB
}
