% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{regressionBasedFeatureSelection}
\alias{regressionBasedFeatureSelection}
\title{select top features via linear regression on an outcome}
\usage{
regressionBasedFeatureSelection(
  dataframein,
  subtypeLabels,
  featureMatrix,
  covariates,
  n_features = 25
)
}
\arguments{
\item{dataframein}{Input dataframe with all relevant data}

\item{subtypeLabels}{Input subtype assignments.}

\item{featureMatrix}{matrix/dataframe defining the data columns as features.}

\item{covariates}{optional string of covariates}

\item{n_features}{select this many features per level}
}
\value{
vector of feature names
}
\description{
select top features via linear regression on an outcome
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
fimp = regressionBasedFeatureSelection( mydf$DX, mydf[,rbfnames] )
}
\author{
Avants BB
}
