% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{reorderingDataframe}
\alias{reorderingDataframe}
\title{Reorder a subtype variable based on an external reference vector}
\usage{
reorderingDataframe(mxdfin, clustername, reorderingVariable)
}
\arguments{
\item{mxdfin}{Input data frame}

\item{clustername}{column name for the identified clusters}

\item{reorderingVariable}{reorder the cluster names based on this variable}
}
\value{
the new dataframe and (if present) membership variables
}
\description{
Reorder a subtype variable based on an external reference vector
}
\author{
Avants BB
}
