% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{replace_values}
\alias{replace_values}
\title{Replace Values in a Vector}
\usage{
replace_values(vec, old_values, new_values)
}
\arguments{
\item{vec}{The original vector in which values need to be replaced.}

\item{old_values}{A vector containing the old values to be replaced.}

\item{new_values}{A vector containing the new values to replace the old ones.}
}
\value{
A vector with the old values replaced by the new values.
}
\description{
This function replaces specified values within a vector with new values.
}
\examples{
replace_values(c(1, 2, 3, 4), c(2, 4), c(20, 40))
}
