% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{replace_values_by_key}
\alias{replace_values_by_key}
\title{Replace Multiple Columns in Dataframe Based on a Matching Key}
\usage{
replace_values_by_key(target_df, source_df, key, columns_to_update)
}
\arguments{
\item{target_df}{A dataframe containing the original values.}

\item{source_df}{A dataframe containing the new values to replace in the target dataframe.}

\item{key}{The column name (as a string) that is used as the key for matching rows in both dataframes.}

\item{columns_to_update}{A character vector of column names in \code{target_df} whose values will be replaced.}
}
\value{
A dataframe with the specified columns' values updated based on the matching key.
}
\description{
This function replaces values in multiple columns of a target dataframe
with values from a source dataframe based on a matching key. The function
takes two dataframes, a key column, and a vector of columns to be updated,
and returns the updated target dataframe.
}
