% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{rowwise_linear_variable_adjustments}
\alias{rowwise_linear_variable_adjustments}
\title{Row-wise linear adjustments to variables in a dataframe}
\usage{
rowwise_linear_variable_adjustments(tempNM, intvars, fixvars, tarints)
}
\arguments{
\item{tempNM}{The input dataframe containing the data to be processed.}

\item{intvars}{A vector of column names in tempNM representing the predictor variables.}

\item{fixvars}{A vector of column names in tempNM representing the variables to be predicted.}

\item{tarints}{A vector of target values corresponding to the predictor variables.}
}
\value{
A dataframe with the predicted values added as new columns. The new columns are named by appending "_pred"
to the original column names specified in fixvars.
}
\description{
Trains linear models on a given dataframe and applies them to predict values.
}
\details{
This function loops through each row in the dataframe, trains a linear model using the predictor values,
and applies the model to predict the values for the specified variables. It handles both simple (one predictor)
and multiple predictor cases, adapting the modeling approach accordingly.

The function uses a simple linear model for prediction, which may not be suitable for complex relationships between variables.
It assumes that the input dataframe has the same structure and column names as expected, and does not perform any error checking
or handling. Therefore, it may fail if the input data is not properly formatted or if there are missing values.
}
\examples{
rowwise_linear_variable_adjustments(mtcars, c("mpg", "cyl"), "disp", c(1, 2))
}
