% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{scale_variables_in_equation}
\alias{scale_variables_in_equation}
\title{Scale Numeric Variables in a Data Frame}
\usage{
scale_variables_in_equation(mydf, myeq, variables_to_exclude)
}
\arguments{
\item{mydf}{A data frame containing the variables to be scaled.}

\item{myeq}{An equation in the form of a string or an object of class
\code{formula} specifying the variables to be scaled. Only the variables
on the right-hand side of the equation are considered.}

\item{variables_to_exclude}{vector of strings not to scale}
}
\value{
A data frame with the specified numeric variables scaled. The
function modifies the input data frame in place and returns it.
}
\description{
This function scales the numeric variables in a data frame that are
specified in a given equation. Non-numeric variables mentioned in the
equation are ignored. Scaling is performed by centering and dividing by the
standard deviation.
}
\examples{
data(iris)
scaled_iris <- scale_variables_in_equation(iris, myeq = "Sepal.Length ~ Sepal.Width + Petal.Length")
head(scaled_iris)

}
