% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{select_important_variables}
\alias{select_important_variables}
\title{Select Important Variables Using Partial Correlation Matrix}
\usage{
select_important_variables(data, cols, threshold = 0.2)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{cols}{A vector of column names or indices to be included in the analysis.}

\item{threshold}{A numeric value between 0 and 1 indicating the proportion of variables to retain. Defaults to 0.2 (20\%).}
}
\value{
A vector of column names corresponding to the most important variables.
}
\description{
This function computes the partial correlation matrix for the specified columns in a dataset and selects the most important variables based on the sum of absolute values of partial correlations.
}
\examples{
set.seed(123)
qqq <- data.frame(matrix(rnorm(100 * 10), ncol = 10))
colnames(qqq) <- paste0("Var", 1:10)
tempcols <- colnames(qqq)
important_vars <- select_important_variables(qqq, tempcols, threshold = 0.3)
print(important_vars)
}
