% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{simlr_impute}
\alias{simlr_impute}
\title{Impute missing SiMLR data in a specified column based on other columns}
\usage{
simlr_impute(dataframe, nms, vecnum, toimpute, family = "gaussian")
}
\arguments{
\item{dataframe}{A data frame containing the data to impute.}

\item{nms}{A vector of base column names.}

\item{vecnum}{A numeric value to append to the column names.}

\item{toimpute}{The base name of the target column to be imputed.}

\item{family}{A string specifying the GLM family (default is 'gaussian').}
}
\value{
A data frame with imputed values.
}
\description{
Impute missing SiMLR data in a specified column based on other columns
}
\examples{
set.seed(123)
expartdfsim <- data.frame(
  t1PC1 = c(1, 2, NA, 4, 5),
  t1aPC1 = c(5, 4, 3, NA, 1),
  dtPC1 = c(2, NA, 3, 4, 5),
  dtaPC1 = c(NA, 3, 2, 1, 5),
  rsfPC1 = c(1, 2, 3, 4, NA),
  perfPC1 = c(NA, 2, 3, 4, 5)
)
nms <- c("t1", "t1a", "dt", "dta", "rsf", "perf")
vecnum <- 1
toimpute <- "perf"
imputed_expartdfsim <- simlr_impute(expartdfsim, nms, vecnum, toimpute, family = 'gaussian')
print(imputed_expartdfsim)
}
