% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{sinkhorn_method}
\alias{sinkhorn_method}
\title{Apply Sinkhorn method to stabilize a correlation matrix}
\usage{
sinkhorn_method(corr_matrix, epsilon = 0.001, max_iter = 100)
}
\arguments{
\item{corr_matrix}{A square correlation matrix to be stabilized.}

\item{epsilon}{Convergence threshold for Sinkhorn iterations (default: 1e-3).}

\item{max_iter}{Maximum number of Sinkhorn iterations (default: 100).}
}
\value{
A Sinkhorn-stabilized correlation matrix.
}
\description{
This function applies the Sinkhorn algorithm to stabilize a given correlation matrix.
}
\examples{
mycor <- abs(cor( matrix(rnorm(100),nrow=10)))
sinkhorn_corr <- sinkhorn_method(mycor)
}
