% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{split_string_vector}
\alias{split_string_vector}
\title{Split a string vector into equal partitions and extract one}
\usage{
split_string_vector(x, n_parts, part_index)
}
\arguments{
\item{x}{Character vector to split.}

\item{n_parts}{Number of equal partitions to split into.}

\item{part_index}{Which partition to return (1-based).}
}
\value{
A character vector corresponding to the requested partition.
}
\description{
Split a string vector into equal partitions and extract one
}
\examples{
sv <- letters[1:20]
split_string_vector(sv, 5, 3)
# returns elements 9:12
}
