% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{table_1_from_formula}
\alias{table_1_from_formula}
\title{Generate a table summarizing the data in a linear model formula.}
\usage{
table_1_from_formula(formula, data, facet_var)
}
\arguments{
\item{formula}{Model formula.}

\item{data}{Data frame containing the data.}

\item{facet_var}{Name of the facet variable.}
}
\value{
A data frame summarizing the data in the formula.
}
\description{
Generate a table summarizing the data in a linear model formula.
}
\examples{
\dontrun{
table_1_from_formula(formula, data, facet_var)
}
}
