% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{topk}
\alias{topk}
\title{Select Top k Rows Based on a Criterion}
\usage{
topk(df, column, k, maximize = TRUE)
}
\arguments{
\item{df}{A dataframe from which to select the top rows.}

\item{column}{A string specifying the column name to base the selection on.}

\item{k}{An integer specifying the number of top rows to select.}

\item{maximize}{A logical value indicating whether to maximize (\code{TRUE}) or
minimize (\code{FALSE}) the criterion. Default is \code{TRUE}.}
}
\value{
A logical vector of the same length as the number of rows in \code{df}.
\code{TRUE} indicates that the row is one of the top \code{k}.
}
\description{
This function selects the top \code{k} rows of a dataframe according to a specified
column and criterion (either maximizing or minimizing the values). It returns
a logical vector indicating which rows are among the top \code{k}.
}
\examples{
df <- data.frame(
  id = 1:10,
  value = c(5, 2, 9, 4, 7, 3, 6, 10, 8, 1)
)

# Select the top 3 rows based on maximizing the 'value' column
topk(df, "value", 3, maximize = TRUE)

# Select the top 3 rows based on minimizing the 'value' column
topk(df, "value", 3, maximize = FALSE)

}
