% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{trainSubtypeUni}
\alias{trainSubtypeUni}
\title{Train subtype for univariate data}
\usage{
trainSubtypeUni(
  mxdfin,
  measureColumn,
  quantiles = c(0.25, 0.75),
  groupVariable,
  group
)
}
\arguments{
\item{mxdfin}{Input data frame}

\item{measureColumn}{string defining a valid column to use for subtyping.  E.g.
the output of the covariate adjustment done by \code{adjustByCovariates}.}

\item{quantiles}{numeric vector defining the quantiles that will yield subgroups}

\item{groupVariable}{names of the column that defines the group to use for training.}

\item{group}{string defining a subgroup on which to train}
}
\value{
data frame defining the subtypes and cutpoints
}
\description{
This is the training module for subtype definition based on a vector.
Produces a data frame denoting the measurement variable, its cutpoints and
the names of the subtypes.
One may want to use a specific sub-group for this, e.g. patients.
}
\examples{
mydf = generateSubtyperData( 100 )
}
\author{
Avants BB
}
