% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{truncatehi}
\alias{truncatehi}
\title{Truncate or Remove High Values}
\usage{
truncatehi(df, x, t = 4, removeit = FALSE)
}
\arguments{
\item{df}{The input data frame.}

\item{x}{The column name to truncate or remove high values from.}

\item{t}{The threshold value (default is 4).}

\item{removeit}{Logical indicating whether to remove or truncate high values (default is FALSE).}
}
\value{
The modified data frame.
}
\description{
Truncates or removes high values in a given column of a data frame.
}
