% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{undersample_majority}
\alias{undersample_majority}
\title{Undersample a majority class in a data frame}
\usage{
undersample_majority(x, vname)
}
\arguments{
\item{x}{data frame containing the data to be undersampled}

\item{vname}{character string specifying the name of the column containing the class labels}
}
\value{
data frame with the majority class undersampled to match the size of the minority class
}
\description{
This function undersamples the majority class in a data frame to match the size of the minority class.
}
\examples{
data <- data.frame(class = c(rep("A", 10), rep("B", 100)), 
                   feature1 = rnorm(110), 
                   feature2 = rnorm(110))
undersampled_data <- undersample_majority(data, "class")
}
