% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{VarSelLCMproba.post}
\alias{VarSelLCMproba.post}
\title{Bug fix to predict function in the VarSelLCM package}
\usage{
VarSelLCMproba.post(object, newdata)
}
\description{
Bug fix to predict function in the VarSelLCM package
}
