% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{visglm}
\alias{visglm}
\title{Visualize the Joint Effect of Variables in a GLM}
\usage{
visglm(
  demogmdl,
  qmdl,
  x,
  y,
  group = NULL,
  titlestring = "",
  varstoadd = NULL,
  groupvar = "group",
  predictorsigns = NULL,
  jdf_simulation = FALSE,
  xrange = NULL,
  verbose = FALSE
)
}
\arguments{
\item{demogmdl}{Data frame containing the variables used in the GLM.}

\item{qmdl}{Fitted GLM object from which predictions will be generated.}

\item{x}{Character vector specifying the names of the predictor variables, with the first being the primary.}

\item{y}{The name of the response variable.}

\item{group}{The name of the group variable or specific group to be analyzed.}

\item{titlestring}{Title of the plot indicating the focus of the visualization.}

\item{varstoadd}{optional additional variables to add into the model; otherwise just extract from the equation}

\item{groupvar}{Optional; the name of the variable in \code{demogmdl} that defines group membership. Default is 'group'.}

\item{predictorsigns}{Optional; a named numeric vector indicating the direction of the effect of each predictor.}

\item{jdf_simulation}{boolean}

\item{xrange}{explicitly set xrange}

\item{verbose}{Logical; if TRUE, additional processing information will be printed to the console.}
}
\value{
Generates a plot visualizing the predicted response and confidence intervals across the range of the primary predictor.
}
\description{
This function visualizes the joint effect of several variables in a generalized linear model (GLM)
by plotting the predicted response over a range of values for the primary predictor variable, while
holding other predictors at their median or mode values. It allows specifying a group for focused analysis.
}
\examples{
# Assuming `data` is your dataset, `fit` is a fitted GLM, 
# and you're interested in predictors `x1` and `x2`:
# visglm(data, fit, c("x1", "x2"), "y", "control", 
#   "Visualization for Control Group")
#
}
