% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{create_table1}
\alias{create_table1}
\title{Create a Table 1 for Population Characteristics}
\usage{
create_table1(
  data,
  summary_vars,
  group_var,
  title = "Table 1: Population Characteristics",
  output_format = "gt",
  landscape = TRUE,
  latex_font_size = c(`12.0pt` = "8.0pt", `14.4pt` = "10.0pt")
)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{summary_vars}{A character vector of variable names to include in the summary.}

\item{group_var}{A character string of the variable name to group by.}

\item{title}{A character string for the table title. Default is "Table 1: Population Characteristics".}

\item{output_format}{A character string indicating the desired output format, either "gt" or "latex". Default is "gt".}

\item{landscape}{boolean}

\item{latex_font_size}{optional font size changes}
}
\value{
A \code{gt} table object or LaTeX code depending on the \code{output_format}.
}
\description{
This function generates a "Table 1" for summarizing population characteristics
in an academic paper. It uses the \code{gtsummary} package for summarizing and
the \code{gt} package for table customization, including adding stripes.
Additionally, if the output format is LaTeX, it generates the corresponding LaTeX code.
}
\examples{
data(mtcars)
mtcars$cyl <- as.factor(mtcars$cyl)
summary_vars <- c("mpg", "hp", "wt")
group_var <- "cyl"
table1_gt <- create_table1(data = mtcars, 
                           summary_vars = summary_vars, 
                           group_var = group_var, 
                           title = "Table 1: Population Characteristics by Cylinder",
                           output_format = "gt")
print(table1_gt)

table1_latex <- create_table1(data = mtcars, 
                               summary_vars = summary_vars, 
                               group_var = group_var, 
                               title = "Table 1: Population Characteristics by Cylinder",
                               output_format = "latex")
cat(table1_latex)
}
