% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{dataPartition}
\alias{dataPartition}
\title{balanced data partition}
\usage{
dataPartition(x, perc, subjectIDs = NULL)
}
\arguments{
\item{x}{target vector to split to train test}

\item{perc}{percentage ( 0 to 1 ) to place in training partition}

\item{subjectIDs}{unique IDs per subject; aids with repeated measurement data
by ensuring subjects exist in uniquely in either train or test}
}
\value{
vector of strings
}
\description{
caret-based data train-test data partition function compatible with mlr3
}
\author{
Avants BB
}
