% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{merge_ppmi_imaging_clinical_demographic_data}
\alias{merge_ppmi_imaging_clinical_demographic_data}
\title{Process Clinical/Demographic and imaging Data for a PPMI Study}
\usage{
merge_ppmi_imaging_clinical_demographic_data(
  demog,
  ppmidemog0,
  pymf,
  pymversion,
  saa,
  verbose = TRUE
)
}
\arguments{
\item{demog}{DataFrame containing PPMI demographic information with columns like PATNO, BIRTHDT.  An example filename from LONI would be Demographics_06Feb2024.csv}

\item{ppmidemog0}{DataFrame containing curated PPMI clinical data including PATNO, EVENT_ID, age_at_visit, age.  An example filename from LONI would be PPMI_Curated_Data_Cut_Public_20230612_rev.csv.}

\item{pymf}{DataFrame containing imaging and additional clinical data keyed by subjectID, date and filename}

\item{pymversion}{A character string identifying the ANTsPyMM pipeline variant.}

\item{saa}{DataFrame (optional) containing supplemental clinical measurements with PATNO, EVENT_ID and CSFSAA.}

\item{verbose}{boolean}
}
\value{
A processed DataFrame with merged and enriched clinical and demographic information.
}
\description{
This function merges, cleans, and enriches clinical and demographic data from a neurological study.
It checks for required columns, replaces placeholder values, merges datasets, and computes new variables.
}
\examples{
\dontrun{
  processed_data <- process_clinical_demographic_data(demog, ppmidemog0, saa, pymf)
}
}
