% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{mlr3classification}
\alias{mlr3classification}
\title{mlr3classifier}
\usage{
mlr3classification(
  dfin,
  tcols,
  learnerName,
  partrate = 0.8,
  dup_size = 0,
  balancing = "smote",
  subjectIDs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dfin}{dataframe input}

\item{tcols}{columns for the prediction task - first is the target outcome}

\item{learnerName}{which mlr3 learner to instantiate}

\item{partrate}{partition ratio for the training 0.8 equals 80 percent train 20 test}

\item{dup_size}{integer for over/under/smote sampling}

\item{balancing}{string over, under, smote, rwo, mwmote, racog, none are the options}

\item{subjectIDs}{unique IDs per subject; aids with repeated measurement data
by ensuring subjects exist in uniquely in either train or test}

\item{verbose}{boolean}
}
\value{
dataframe with task, learner, accuracy and balanced accuracy
}
\description{
build a mlr3 classification model
}
\author{
Avants BB
}
