% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{normativeSummary}
\alias{normativeSummary}
\title{Normative Summary}
\usage{
normativeSummary(
  data,
  subjectRow,
  columns,
  zoom,
  idcolumn = "commonID",
  sexcol = "commonSex",
  agecol = "commonAge",
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame containing the dataset for analysis.}

\item{subjectRow}{The row number or identifier for the subject of interest within the dataset.}

\item{columns}{A vector of column names in the dataset that should be summarized.}

\item{zoom}{A parameter to specify the focus or zoom level of the summary.  zoom (integer) specifies the number of nearest neighbors to use.}

\item{idcolumn}{column name for the unique subject ID}

\item{sexcol}{name of the sex column}

\item{agecol}{name of the age column}

\item{verbose}{A logical flag indicating whether to print detailed output.}
}
\value{
A summary object containing normative statistics for the specified columns of the subject.
}
\description{
This function provides a normative summary for a given subject within a dataset. It can be configured
to focus on specific columns and can be zoomed into particular data points.
}
\examples{
# normativeSummary(myData, 1, c("Column1", "Column2"), zoom = 1, verbose = TRUE)
}
