% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{outlierness}
\alias{outlierness}
\title{Get subjects and timepoints with the best quality}
\usage{
outlierness(mxdfin, measureColumns, calck, report = FALSE)
}
\arguments{
\item{mxdfin}{Input data frame with repeated measurements and a grouped time variable}

\item{measureColumns}{vector string of column names to use in outlier detection}

\item{calck}{optional integer for knn}

\item{report}{boolean return vector string of column names for outlier detection results}
}
\value{
data frame
}
\description{
This filters data that has both repeats (e.g. multiple images on the same day)
and (optionally) longitudinal data.  The quality measure should be a "higher"
is better" measurement.
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
mydfol = outlierness( mydf, rbfnames )
}
\author{
Avants BB
}
