% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{prplot}
\alias{prplot}
\title{prplot}
\usage{
prplot(
  mdl,
  xvariable,
  byvariable,
  titlestring = "",
  ystring = "",
  addpoints = 0,
  palette = "npg"
)
}
\arguments{
\item{mdl}{input fitted model}

\item{xvariable}{a model term}

\item{byvariable}{a model term interacting with the xvariable}

\item{titlestring}{string for title}

\item{ystring}{string for title}

\item{addpoints}{continuous value greater than zero}

\item{palette}{string}
}
\value{
the quantile transformed vector
}
\description{
partial residual regression plot using ggpubr for display
and visreg for partial residual calculation
}
\author{
Avants BB
}
