% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{rank_methods_by_performance}
\alias{rank_methods_by_performance}
\title{Rank Rows of a Data Frame by Weighted Performance}
\usage{
rank_methods_by_performance(df, id_col, weights_df)
}
\arguments{
\item{df}{The performance data frame.}

\item{id_col}{A character string specifying the unique identifier column.}

\item{weights_df}{A data frame defining the metrics, weights, and directions.}
}
\value{
A list containing \code{ranked_df} (a tibble) and \code{gt_table} (a gt object).
}
\description{
Rank Rows of a Data Frame by Weighted Performance
}
