% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{regressionBasedFeatureSelection}
\alias{regressionBasedFeatureSelection}
\title{select top features via linear regression on an outcome}
\usage{
regressionBasedFeatureSelection(
  dataframein,
  subtypeLabels,
  featureMatrix,
  covariates = "1",
  n_features = 25,
  associationType = "features2subtypes"
)
}
\arguments{
\item{dataframein}{Input dataframe with all relevant data}

\item{subtypeLabels}{Input subtype assignments.}

\item{featureMatrix}{matrix/dataframe defining the data columns as features.}

\item{covariates}{optional string of covariates}

\item{n_features}{select this many features per level}

\item{associationType}{either predictor features from subtypes or predict
subtypes from features.  will produce related but complementary results. in
some cases, depending on subtypes/degrees of freedom, only one will be appropriate.}
}
\value{
vector of feature names
}
\description{
select top features via linear regression on an outcome
}
\examples{
mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
fimp = regressionBasedFeatureSelection( mydf, mydf$DX, mydf[,rbfnames], 
   associationType ="subtypes2features" )
}
\author{
Avants BB
}
