% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{run_association_analysis}
\alias{run_association_analysis}
\title{A master function to run association analyses (LM or LMER).}
\usage{
run_association_analysis(
  data,
  predictors,
  outcomes,
  covariates,
  dataset_name,
  analysis_type = "Cross-Sectional",
  use_delta_outcome = TRUE,
  random_effects = "(1 | PTID)"
)
}
\arguments{
\item{data}{The dataframe containing all variables.}

\item{predictors}{A character vector of predictor column names.}

\item{outcomes}{A character vector of outcome column names.}

\item{covariates}{A character string of covariates.}

\item{dataset_name}{A string to label the results.}

\item{analysis_type}{A string, either "Cross-Sectional" or "Longitudinal".}

\item{use_delta_outcome}{A logical value. If TRUE (default), appends "_delta"
to the outcome variable for longitudinal analyses.}

\item{random_effects}{A string for the random effects part of the LMER formula.}
}
\value{
A tidy data frame with results for every predictor-outcome pair.
}
\description{
This function acts as a dispatcher, calling the appropriate helper
(lm_anv_p_and_d or lmer_anv_p_and_d) based on whether the analysis
is specified as longitudinal. It extracts and returns key metrics.
}
