% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{threewayinteraction}
\alias{threewayinteraction}
\title{three way interaction plot from raw data}
\usage{
threewayinteraction(
  indf,
  xvar,
  yvar,
  colorvar,
  anat,
  anatshow,
  pallete = c("#00AFBB", "#E7B800", "#FC4E07")
)
}
\arguments{
\item{indf}{dataframe with relevant variables}

\item{xvar}{variable for x-axis of plots}

\item{yvar}{variable for y-axis of plots}

\item{colorvar}{variable by which to color plots}

\item{anat}{continuous variable by which to split plots}

\item{anatshow}{character name for continuous variable to show on plots (optional)}

\item{palette}{optional palette}
}
\value{
the plot
}
\description{
three way interaction plot from raw data
}
\examples{
# FIXME
}
\author{
Avants BB
}
