% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtyper.R
\name{threewayinteraction}
\alias{threewayinteraction}
\title{three way interaction plot from raw data}
\usage{
threewayinteraction(
  indf,
  xvar,
  yvar,
  colorvar,
  anat,
  anatshow,
  ggpalette = "jco",
  showpoints = FALSE
)
}
\arguments{
\item{indf}{dataframe with relevant variables}

\item{xvar}{variable for x-axis of plots}

\item{yvar}{variable for y-axis of plots}

\item{colorvar}{variable by which to color plots}

\item{anat}{continuous variable by which to split plots}

\item{anatshow}{optional character name for continuous variable to show on plots - can be up to vector length 3 - one for each panel of the plot}

\item{ggpalette}{optional palette}

\item{showpoints}{boolean}
}
\value{
the plot
}
\description{
three way interaction plot from raw data
}
\examples{
# FIXME
}
\author{
Avants BB
}
